feat: completed the task of the subtask "Modules(src/modules)"
