+++ src/modules/files/c.js
@@ -1,1 +1,4 @@
-console.log('Hello from c.js!');
+ const info = () => {
+    console.log('Hello from c.js!');
+}
+ export default info();
